refactor(server): split routes into books vs accounting concerns
Introduce a structural books/accounting boundary with no behavior change.
Books writes facts (ingest, categorize, journal); accounting derives
meaning (chart of accounts, reporting, tax, allocations).

- move transactions/import/webhooks -> server/books/
- move accounts/reports/tax/allocations -> server/accounting/
- add concern-boundary READMEs for each tree
- classification.ts stays in routes/ (straddles both); note added
- fix import paths in app.ts and __tests__
