+++ server/accounting/README.md
@@ -1,0 +1,7 @@
+# Accounting
+
+Accounting **derives meaning** from the facts Books records.
+
+Boundary: chart of accounts, reporting, tax, and allocations. It consumes the facts written by `server/books/` (transactions, imports, webhooks) and does not ingest or journal raw activity itself.
+
+Owner: `chittyaccounting-agent` → coa / reporting / tax / allocations.
