+++ server/app.ts
@@ -10,15 +10,15 @@ import { callerContext } from './middleware/caller';
 import { tenantMiddleware } from './middleware/tenant';
 import { healthRoutes } from './routes/health';
 import { docRoutes } from './routes/docs';
-import { accountRoutes } from './routes/accounts';
+import { accountRoutes } from './accounting/accounts';
 import { summaryRoutes } from './routes/summary';
 import { tenantRoutes } from './routes/tenants';
 import { propertyRoutes } from './routes/properties';
-import { transactionRoutes } from './routes/transactions';
+import { transactionRoutes } from './books/transactions';
 import { integrationRoutes } from './routes/integrations';
 import { taskRoutes } from './routes/tasks';
 import { aiRoutes } from './routes/ai';
-import { webhookRoutes } from './routes/webhooks';
+import { webhookRoutes } from './books/webhooks';
 import { mercuryRoutes } from './routes/mercury';
 import { githubRoutes } from './routes/github';
 import { stripeRoutes } from './routes/stripe';
@@ -27,16 +27,16 @@ import { chargeRoutes } from './routes/charges';
 import { forensicRoutes } from './routes/forensics';
 import { valuationRoutes } from './routes/valuation';
 import { portfolioRoutes } from './routes/portfolio';
-import { importRoutes } from './routes/import';
+import { importRoutes } from './books/import';
 import { mcpRoutes } from './routes/mcp';
-import { reportRoutes } from './routes/reports';
-import { taxRoutes } from './routes/tax';
+import { reportRoutes } from './accounting/reports';
+import { taxRoutes } from './accounting/tax';
 import { googleRoutes, googleCallbackRoute } from './routes/google';
 import { commsRoutes } from './routes/comms';
 import { workflowRoutes } from './routes/workflows';
 import { leaseRoutes } from './routes/leases';
 import { chittyIdAuthRoutes } from './routes/chittyid-auth';
-import { allocationRoutes } from './routes/allocations';
+import { allocationRoutes } from './accounting/allocations';
 import { classificationRoutes } from './routes/classification';
 import { emailRoutes } from './routes/email';
 import { createDb } from './db/connection';
